chore: separating modal tabs into separate files; feat: eligibility and benefits tab
